+++ platformio/builder/tools/platformio.py
@@ -9,8 +9,19 @@ from os.path import basename, dirname, isdir, isfile, join, normpath
 from SCons.Script import Exit, SConscript, SConscriptChdir
 from SCons.Util import case_sensitive_suffixes
 
+from platformio import __version__
+
 
 def BuildFirmware(env):
+
+    def _append_pioversion():
+        vermatch = re.match(r"^([\d\.]+)", __version__)
+        assert vermatch
+        intparts = [int(i) for i in vermatch.group(1).split(".")[:3]]
+        firmenv.Append(
+            CPPDEFINES=["PLATFORMIO={:02d}{:02d}{:02d}".format(*intparts)]
+        )
+
     # fix ASM handling under non-casitive OS
     if not case_sensitive_suffixes('.s', '.S'):
         env.Replace(
@@ -51,6 +62,8 @@ def BuildFirmware(env):
                              env.subst("$SRCBUILD_FLAGS"))
     if _srcbuild_flags:
         firmenv.MergeFlags(_srcbuild_flags)
+
+    _append_pioversion()
 
     return firmenv.Program(
         join("$BUILD_DIR", "firmware"),